Move context out of components folder and reduce other circular dependencies
