+++ packages/react-querybuilder/src/components/index.ts
@@ -3,10 +3,8 @@ export * from './DragHandle';
 export * from './InlineCombinator';
 export * from './NotToggle';
 export * from './QueryBuilder';
-export * from './QueryBuilderContext';
 export * from './Rule';
 export * from './RuleGroup';
 export * from './ShiftActions';
 export * from './ValueEditor';
 export * from './ValueSelector';
-export * from './defaults';
