+++ packages/react-querybuilder/src/hooks/useRuleGroup.ts
@@ -1,7 +1,6 @@
 import type { MouseEvent } from 'react';
 import { useCallback, useMemo } from 'react';
 import { standardClassnames } from '../defaults';
-import { useDeprecatedProps, useReactDndWarning } from '../hooks';
 import type {
   ActionElementEventHandler,
   Classnames,
@@ -21,6 +20,8 @@ import {
   pathsAreEqual,
 } from '../utils';
 import { clsx } from '../utils/clsx';
+import { useDeprecatedProps } from './useDeprecatedProps';
+import { useReactDndWarning } from './useReactDndWarning';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export type UseRuleGroup = Omit<RuleGroupProps, 'ruleGroup'> & {
