+++ packages/react-querybuilder/src/index.ts
@@ -1,5 +1,6 @@
 export * from './components';
 export { QueryBuilder as default } from './components';
+export * from './defaultControlElements';
 export * from './defaults';
 export * from './hooks';
 export * from './messages';
