+++ packages/react-querybuilder/src/utils/index.ts
@@ -23,7 +23,6 @@ export * from './regenerateIDs';
 export * from './toFullOption';
 export * from './toOptions';
 export * from './transformQuery';
-export * from './uniq';
 
 // Don't export clsx. It should be imported from the official clsx
 // package if used outside the context of this package.
